# nodejs/swagger/auth.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documenting API UTN Ferrelepe with Swagger.",
        "version": "0.2.0",
        "title": "API Documented"
    },
    "host": "localhost:8000",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/users/user": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create User",
                "description": "",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [{
                        "name": "name",
                        "in": "formData",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Phone",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "formData",
                        "description": "State of user",
                        "required": true,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user was created successfully"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [{
                    "auth_token": [
                        "write:user",
                        "read:user"
                    ]
                }]
            },
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User By Token",
                "description": "",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "User successfully found"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [{
                    "auth_token": [
                        "write:user",
                        "read:user"
                    ]
                }]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit user by ID",
                "description": "",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],

                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Phone",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully modified"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [{
                    "auth_token": [
                        "write:user",
                        "read:user"
                    ]
                }]
            }
        },
        "/users/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return a user by ID",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "User successfully found"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [{
                    "auth_token": [
                        "write:user",
                        "read:user"
                    ]
                }]
            }
        },
...
    },
    "securityDefinitions": {
        "auth_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}